refactor: Rename applyTree to treeUtil
