+++ src/AlisEditor.vue
@@ -38,12 +38,12 @@ import Vue from 'vue'
 import uuid from 'uuid/v4'
 import EditorBlock from './components/blocks/EditorBlock.vue'
 import EditorToolbar from './components/menu/EditorToolbar.vue'
-import InsertPopup from './components/utils/InsertPopup.vue'
+import InsertPopup from './components/menu/InsertPopup.vue'
 import { Block, BlockType, ParagraphBlock } from './types/Blocks'
 import { createBlock, isContentEditableBlock } from './utils/createBlock'
 import { createDataURIImage } from './utils/createImage'
 import { isMobile, isDesktop } from './utils/deviceUtil'
-import { findRootIdByBlockId, findTreeContentById, findBeforeRootContentByRootBlockId } from './utils/applyTree'
+import { findRootIdByBlockId, findTreeContentById, findBeforeRootContentByRootBlockId } from './utils/treeUtil'
 import { EditorStore } from './store/'
 import { cloneDeep } from 'lodash'
 import urlregex from 'url-regex'
