+++ tests/unit/utils/treeUtil.ts
@@ -1,8 +1,8 @@
 import { BlockType } from '../../../src/types/Blocks'
 import { createBlock } from '../../../src/utils/createBlock'
-import * as applyTree from '../../../src/utils/applyTree'
+import * as treeUtil from '../../../src/utils/treeUtil'
 
-describe('applyTree.ts', () => {
+describe('treeUtil.ts', () => {
   describe('.findBeforeRootContentByRootBlockId', () => {
     describe(`('456', [{id: '123'}, {id:'456'}])`, () => {
       it('shoud returns "123"', () => {
@@ -10,7 +10,7 @@ describe('applyTree.ts', () => {
           createBlock(BlockType.Paragraph, { id: '123' }),
           createBlock(BlockType.Paragraph, { id: '456' })
         ]
-        expect(applyTree.findBeforeRootContentByRootBlockId('456', blocks)).toMatchObject(
+        expect(treeUtil.findBeforeRootContentByRootBlockId('456', blocks)).toMatchObject(
           createBlock(BlockType.Paragraph, { id: '123' })
         )
       })
@@ -22,13 +22,13 @@ describe('applyTree.ts', () => {
           createBlock(BlockType.Paragraph, { id: '123' }),
           createBlock(BlockType.Paragraph, { id: '456' })
         ]
-        expect(applyTree.findBeforeRootContentByRootBlockId('1000', blocks)).toBe(null)
+        expect(treeUtil.findBeforeRootContentByRootBlockId('1000', blocks)).toBe(null)
       })
     })
   })
   describe(`.findRootIdByBlockId('456',  [{id:'123',children:[{id:'456'}]}])`, () => {
     it('should returns "123"', () => {
-      expect(applyTree.findRootIdByBlockId('456', [{ id: '123', children: [{ id: '456' }] }] as any)).toBe('123')
+      expect(treeUtil.findRootIdByBlockId('456', [{ id: '123', children: [{ id: '456' }] }] as any)).toBe('123')
     })
   })
 })
